+++ JobFinder/Model/JobApplication/Params/GetJobApplicationByUserIdParam.cs
@@ -1,0 +1,11 @@
+using JobFinder.Model.Utils.Fetching;
+using JobFinder.Model.Utils.Fetching.Filters;
+
+namespace JobFinder.Model;
+
+public class GetJobApplicationByUserIdParam
+{
+    public GetJobApplicationByUserIdFilter? Filter { get; set; }
+    public Order? Order { get; set; }
+    public Pagination? Pagination { get; set; }
+}
