+++ JobFinder/Model/Utils/Fetching/Filters/Company/CompanyFilter.cs
