+++ JobFinder/Model/Utils/Fetching/Filters/Job/JobFilter.cs
