+++ JobFinder/Model/Utils/Fetching/Filters/JobApplication/JobApplicationFilter.cs
